requirements.txt
streamlit=1.35.0
pandas=2.2.2
openai=1.30.3
python-dotenv=1.0.1
openpyxl=3.1.2
streamlit-chat=0.1.1